refactor: move some func to utils file
