+++ src/constants.ts
@@ -1,0 +1,9 @@
+export const DEFAULT_PORT = 4000;
+
+export const YAML_SWAGGER_FILENAME = 'api.yaml';
+
+export const PATH_TO_YAML_FOLDER = './doc';
+
+export enum Messages {
+  OLD_PASSWORD_WRONG = 'Old password is wrong',
+}
