+++ src/utils.ts
@@ -1,0 +1,7 @@
+import { EntityType } from './types';
+
+export const getNotFoundMessage = (type: EntityType, id: string) => {
+  return `${
+    type.charAt(0).toUpperCase() + type.slice(1)
+  } with id '${id}' does not exist`;
+};
